[step4] refactor 패키지 변경
