+++ src/main/java/model/lotto/LottoNo.java
@@ -1,4 +1,4 @@
-package model;
+package model.lotto;
 
 import java.util.HashMap;
 import java.util.Map;
