+++ src/main/java/model/lotto/ManualLotto.java
@@ -1,4 +1,4 @@
-package model;
+package model.lotto;
 
 public class ManualLotto extends Lotto {
